Move MysqlAlterColumn to renderer
